+++ examples/admin-partitions/terraform/variables.tf
@@ -18,7 +18,7 @@ variable "consul_image" {
 variable "consul_ecs_image" {
   description = "Consul ECS image to use."
   type        = string
-  default     = "public.ecr.aws/hashicorp/consul-ecs:0.4.1"
+  default     = "public.ecr.aws/hashicorp/consul-ecs:0.5.0-beta1"
 }
 
 variable "client_partition" {