+++ modules/mesh-task/main.tf
@@ -2,7 +2,7 @@ data "aws_region" "current" {}
 
 locals {
   // Must be updated for each release, and after each release to return to a "-dev" version.
-  version_string = "0.4.1-dev"
+  version_string = "0.5.0-dev"
 
   gossip_encryption_enabled = var.gossip_key_secret_arn != ""
   consul_data_volume_name   = "consul_data"
@@ -24,7 +24,9 @@ locals {
     "/consul/consul-ecs", "app-entrypoint", "-shutdown-delay", "${var.application_shutdown_delay_seconds}s",
   ] : null
   app_mountpoints = local.enable_app_entrypoint ? [local.consul_data_mount] : []
-  service_name    = var.consul_service_name != "" ? var.consul_service_name : var.family
+
+  // Lower case service name is required. var.consul_service_name is validated to be lower case, while the task family is forced to lower case
+  service_name = var.consul_service_name != "" ? var.consul_service_name : lower(var.family)
 
   // Optionally, users can provide a partition and namespace for the service.
   partition_tag = var.consul_partition != "" ? { "consul.hashicorp.com/partition" = var.consul_partition } : {}